Split huge testfile into multiple files Add test for AddAnnotation and DeleteAnnotation.
