+++ packages/test/src/test/delta/ReferenceCommands.test.ts
@@ -1,0 +1,29 @@
+import { RepositoryClient } from "@lionweb/server-http-client"
+import { adminResponseFunctions, DeltaClient, eventFunctions, responseFunctions } from "@lionweb/server-delta-client"
+import { GetAvailableIdsResponse, ListPartitionsResponse } from "@lionweb/server-delta-shared"
+import { HttpSuccessCodes } from "@lionweb/server-shared"
+import { test, describe, beforeAll, beforeEach, afterAll } from "vitest"
+import { LionWebTreeConverter } from "../models/LionWebTree.js"
+import { LibraryModel, libraryNodes, LibraryTree, ProgramModel, programNodes, ProgramTree, resetModels } from "../models/testmodel.js"
+import { ProcedureBody, ProgramCommands } from "../models/keys.js"
+import { CoverageMap, cmd, expectError, expectEvent, expectResponse, logProtocol } from "./test-helpers.js"
+import { client, checkClient, logoModel, beforeAllTests, bulkApiClient, log, afterAllTests, withoutHistoryList } from "./SharedTest.js"
+
+describe.each(withoutHistoryList)("Delta tests without history: $withoutHistory ", async ({ withoutHistory }) => {
+    beforeAll(async function () {
+        await beforeAllTests(withoutHistory)
+    })
+
+    afterAll(async function () {
+        afterAllTests()
+    })
+
+    beforeEach(async function () {
+        client.sentMessageHistory = []
+        client.receivedMessageHistory = []
+    })
+
+    test("dummy", () => {
+        
+    })
+})
